refactor: rewrite structure of the project
