--- src/blocks/not-found/__error/not-found__error.css
@@ -1,15 +1,0 @@
-.not-found__error {
-  margin-top: 39px;
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 19px;
-}
-
-@media screen and (max-width: 400px) {
-  .not-found__error {
-    font-size: 12px;
-    line-height: 15px;
-  }
-}
