--- src/blocks/not-found/not-found.css
@@ -1,7 +1,0 @@
-.not-found {
-  display: flex;
-  max-width: 1280px;
-  margin: 0 auto;
-  flex-direction: column;
-  text-align: center;
-}
