+++ src/components/Login/Login.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from '../images/logo.svg';
+import logo from '../../images/logo.svg';
 
 function Login(props) {
   return (
